% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfda_rainbow.R
\name{as.fds}
\alias{as.fds}
\alias{as.fds.npf.data}
\title{Converts functional data to a fds object}
\usage{
as.fds(obj, ...)

\method{as.fds}{npf.data}(obj, ...)
}
\arguments{
\item{obj}{a functional data object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns a \link[rainbow:fds]{rainbow::fds}-class object.
}
\description{
Converts a functional data object to a \link[rainbow:fds]{fds} object.
}
\examples{
fdata <- npf.data(ozone, dimnames = "day")
rainbow::plot.fds(as.fds(fdata))
}
