% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.cv.npf.R
\name{h.cv.npf.bin}
\alias{h.cv.npf.bin}
\alias{h.cv.npf}
\alias{h.cv.npf.bin.res2}
\title{Cross-validation methods for bandwidth selection}
\usage{
\method{h.cv}{npf.bin}(bin, var, cor, ...)

\method{h.cv}{npf.bin.res2}(bin, var, cor, ...)
}
\arguments{
\item{bin}{object used to select a method (binned data).}

\item{var}{(optional) a vector or an object of class \code{\link{npf.var}} with the estimated (or
theoretical) variances.}

\item{cor}{(optional) intra-curve correlation matrix or semivariogram model
(\code{\link{svarmod}}-class) of the data. Defaults to the identity matrix.}

\item{...}{further arguments passed to \code{\link[npsp:h.cv]{npsp::h.cv}}.}
}
\description{
Selects the bandwidth of a local polynomial kernel (regression or
variance) estimator using (standard or modified) CV, GCV or MASE criteria.
See \code{\link[npsp:h.cv]{npsp::h.cv}} for details.
}
\seealso{
\code{\link[npsp:h.cv]{npsp::h.cv}}, \code{\link{npf.binning}}, \code{\link{npf.bin.res2}}
}
