% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locpol.npf.R
\name{locpol.npf.data}
\alias{locpol.npf.data}
\alias{npf.locpol-class}
\alias{npf.locpol}
\alias{locpol.npf.bin}
\alias{predict.npf.locpol}
\title{Local polynomial trend estimation}
\usage{
\method{locpol}{npf.data}(
  x,
  h = NULL,
  degree = 1 + as.numeric(drv),
  drv = FALSE,
  hat.bin = FALSE,
  ncv = 0,
  ...
)

\method{locpol}{npf.bin}(
  x,
  h = NULL,
  degree = 1 + as.numeric(drv),
  drv = FALSE,
  hat.bin = FALSE,
  ncv = 0,
  ...
)

\method{predict}{npf.locpol}(object, newx = NULL, ...)
}
\arguments{
\item{x}{an object (of class \code{\link{npf.data}} or \code{\link{npf.bin}}) used to select a method.}

\item{h}{(full) bandwidth matrix (controls the degree of smoothing;
only the upper triangular part of h is used).}

\item{degree}{degree of the local polynomial used. Defaults to 1 (local linear estimation).}

\item{drv}{logical; if \code{TRUE}, the matrix of estimated first derivatives is returned.}

\item{hat.bin}{logical; if \code{TRUE}, the hat matrix of the binned data is returned.}

\item{ncv}{integer; determines the number of cells leaved out in each dimension.
Defaults to 0 (the full data is used) and it is not normally changed by the user
in this setting. See "Details" below.}

\item{...}{further arguments passed to or from other methods.}

\item{object}{object used to select a method.}

\item{newx}{vector with the (irregular) points to predict (interpolate).}
}
\value{
\code{locpol.npf.data} and  \code{locpol.npf.bin} return an S3 object of class
\code{npf.locpol} extending \code{\link[npsp:locpol]{npsp::locpol.bin}} (locpol + bin data + grid par.).

If \code{newx == NULL}, \code{predict.npf.locpol} returns the trend estimates
corresponding to the discretization points
(otherwise \code{npsp::interp.data.grid} is called).
}
\description{
Estimates the functional mean (and its first derivatives)
using local polynomial kernel smoothing.
}
\examples{
fd <- npf.data(ozone, dimnames = "day")
# Linear Local trend estimate
lp <- locpol(fd, h = 35)
# Plot
plot(fd, col = "lightgray", legend = FALSE)
lines(lp$data$x, lp$biny, lty = 2) # x = coords(fd)
lines(lp$data$x, lp$est)
# Bandwidth selection
# (assuming independence and homoscedasticity)
bin <- npf.binning(fd) # binning
trend.h <- h.cv(bin)$h
trend.h
# the selected bandwidth undersmoothes the data...
trend.h <- 4*trend.h
# Linear Local trend estimate
lp <- locpol(bin, h = trend.h, hat.bin = TRUE)
# Plot
plot(fd, col = "lightgray", legend = FALSE)
lines(lp$data$x, lp$biny, lty = 2)
lines(lp$data$x, lp$est)
}
\seealso{
\code{\link[npsp:locpol]{npsp::locpol.bin}}, \code{\link{npf.bin}}
}
