% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.svar.npf.R
\name{np.svar.npf.data}
\alias{np.svar.npf.data}
\alias{np.svar.npf.locpol}
\title{Local polynomial estimation of the (intra-curve) semivariogram}
\usage{
\method{np.svar}{npf.data}(
  x,
  h = NULL,
  maxlag = NULL,
  nlags = round(0.55 * x$grid$n),
  minlag = x$grid$lag,
  degree = 1,
  drv = FALSE,
  hat.bin = TRUE,
  ncv = 0,
  ...
)

\method{np.svar}{npf.locpol}(
  x,
  var,
  h = NULL,
  maxlag = NULL,
  nlags = round(0.55 * x$grid$n),
  minlag = x$grid$lag,
  degree = 1,
  drv = FALSE,
  hat.bin = TRUE,
  ncv = 0,
  ...
)
}
\arguments{
\item{x}{object used to select a method.}

\item{h}{(full) bandwidth matrix (controls the degree of smoothing;
only the upper triangular part of h is used).}

\item{maxlag}{maximum lag. Defaults to \code{(nlags - 1) * minlag}.}

\item{nlags}{number of lags. Defaults to \code{round(0.55 * x$grid$n)}.}

\item{minlag}{minimun lag. Defaults to that of the discretization grid.}

\item{degree}{degree of the local polynomial used. Defaults to 1 (local linear estimation).}

\item{drv}{logical; if \code{TRUE}, the matrix of estimated first derivatives is returned.}

\item{hat.bin}{logical; if \code{TRUE}, the hat matrix of the binned data is returned.}

\item{ncv}{integer; determines the number of cells leaved out in each dimension.
Defaults to 0 (the full data is used) and it is not normally changed by the user
in this setting. See "Details" below.}

\item{...}{further arguments passed to \code{\link[npsp:svar.bin]{npsp::svariso}}.}

\item{var}{(optional) a vector or an object of class \code{\link{npf.var}} with the estimated (or
theoretical) variances.}
}
\value{
Returns an object of S3 class \code{\link[npsp:np.svar]{npsp::np.svar}}.
}
\description{
Creates a \code{svar.bin} (binned semivar. + grid parameters) object with
the binned intra-curves semivariances (i.e. computes a binned sample variogram).
}
\details{
Only one of the arguments \code{maxlag} or \code{nlags} must be specified.

See \code{\link[npsp:np.svar]{npsp::np.svar}} for more details.
}
\seealso{
\code{\link[npsp:np.svar]{npsp::np.svar}}, \code{\link{svar.bin.npf.data}}, \code{\link{svar.bin.npf.locpol}}
}
