% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.var.npf.R
\name{npf.bin.res2}
\alias{npf.bin.res2}
\alias{npf.bin.res2-class}
\title{Binning of squared residuals of a local polynomial fit}
\usage{
npf.bin.res2(x)
}
\arguments{
\item{x}{local polynomial fit (\code{\link{npf.locpol}}-\code{\link{class}} object).}
}
\value{
Returns an S3 object of class \code{npf.bin.res2} extending
\code{\link[npsp:binning]{npsp::bin.data}} (bin data + grid par.).
}
\description{
Aggregates the squared residuals of a local polynomial kernel smoothing
into the regular grid of discretization points,
i.e. computes the functional residual variance (using \eqn{n} as
denominator) and the sample size.
}
\examples{
fd <- npf.data(ozone, dimnames = "day")
# Linear Local trend estimate
lp <- locpol(fd, h = 35)
# Bandwidth selection for variance estimation
bin.res2 <- npf.bin.res2(lp)
h.cv(bin.res2)
}
\seealso{
\code{\link{npf.binning}}, \code{\link{npf.var}}
}
