% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locpol.npf.R
\name{npf.binning}
\alias{npf.binning}
\alias{npf.bin-class}
\alias{npf.bin}
\title{Binning of functional data}
\usage{
npf.binning(x)
}
\arguments{
\item{x}{funcional data (\code{\link{npf.data}}-\code{\link{class}} object).}
}
\value{
Returns an S3 object of class \code{npf.bin} extending
\code{\link[npsp:binning]{npsp::bin.data}} (bin data + grid par.).
}
\description{
Aggregates the functional data into the regular grid of discretization points,
i.e. computes the functional sample mean and the sample size.
}
\examples{
fd <- npf.data(ozone, dimnames = "day")
bin <- npf.binning(fd)
plot(fd, col = "lightgray", legend = FALSE)
with(bin, lines(data$x, biny))
}
\seealso{
\code{\link[npsp:binning]{npsp::bin.data}}, \code{\link{locpol.npf.bin}}
}
