% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npf.data.R
\name{npf.data}
\alias{npf.data}
\alias{npf.data-class}
\alias{coords.npf.data}
\title{Functional data (S3 class "npf.data")}
\usage{
npf.data(y, min, max, lag, ylabel, ynames, dimnames)

\method{coords}{npf.data}(x, ...)
}
\arguments{
\item{y}{numeric matrix or array; the last dimension corresponds to curves and
the first ones to discretization points.}

\item{min}{vector; minimum values of the discretization points (coordinates).}

\item{max}{vector; maximum values of the discretization points (optional).}

\item{lag}{vector; lag in each dimension (optional).}

\item{ylabel}{character; name used to label the data.}

\item{ynames}{character vector; names used to label the curves.}

\item{dimnames}{character vector; names used to label the dimensions.}

\item{x}{funcional data (\link{npf.data}-\link{class} object).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns an object of \code{\link{class}} \code{npf.data}
(extending \code{npsp::data.grid}), a list with
the arguments \code{y}, \code{ylabel} and \code{ynames} as components, and two additional
components:
\itemize{
\item \code{ny} number of curves (sample size).
\item \code{grid} a \link[npsp:grid.par]{npsp::grid.par}-class object that defines the
discretization points.
}
}
\description{
Defines a multidimensional functional data set.
Constructor function of the \code{npf.data}-\code{\link{class}}.
}
\details{
Only one of the arguments \code{max} or \code{lag} must be specified.
}
\examples{
fd.ozone <- npf.data(ozone, dimnames = "day")
plot(fd.ozone)
}
\seealso{
\code{\link[npsp:data.grid]{npsp::data.grid}}.
}
