% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npf.fit.R
\name{npf.fit}
\alias{npf.fit}
\alias{npf.fit.default}
\alias{npf.fit.npf.model}
\title{Fit a nonparametric functional model}
\usage{
npf.fit(x, ...)

\method{npf.fit}{default}(
  x,
  iter = 2,
  tol = 0.1,
  h = NULL,
  var.h = NULL,
  svar.h = NULL,
  maxlag = NULL,
  nlags,
  dk = 0,
  verbose = FALSE,
  ...
)

\method{npf.fit}{npf.model}(
  x,
  iter = 1,
  tol = 0.1,
  svar.h = x$svm$esv$locpol$h,
  dk = x$svm$par$dk,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{a (data) object used to select a method.}

\item{...}{further arguments passed to \code{\link{h.cv.npf.bin}}
(trend bandwidth selection parameters).}

\item{iter}{maximum number of iterations.}

\item{tol}{relative convergence tolerance (trend and variance bandwidths).}

\item{h}{initial bandwidth for trend estimation
(final bandwidth if \code{iter = 0}).}

\item{var.h}{initial bandwidth for variance estimation
(final bandwidth if \code{iter = 0}).}

\item{svar.h}{bandwidth for variogram estimation.}

\item{maxlag, nlags}{maximum lag and number of lags in variogram estimation
(only one must be specified; see \code{\link{np.svar.npf.locpol}}).}

\item{dk}{dimension of the Shapiro-Botha variogram model (see \code{\link[npsp:fitsvar.sb.iso]{npsp::fitsvar.sb.iso}}).}

\item{verbose}{logical; if \code{TRUE}, the errors (the relative differences
in trend and variance bandwidth selections) at each iteration are printed.}
}
\value{
Returns an object of class \code{npf.model}
(extending \code{\link{npf.locpol}}), the \code{lp} argument with the others as additional
components, and an additional component \code{convergence} with the number of iterations and
the relative differences in trend and variance bandwidth selections.
}
\description{
Fits a nonparametric functional model
(jointly estimates the trend, the variance and the variogram) by calling
\code{\link{locpol.npf.bin}}, \code{\link{np.var.npf.bin.res2}}, \code{\link{np.svar.npf.locpol}} and
\code{\link[npsp:fitsvar.sb.iso]{npsp::fitsvar.sb.iso}} iteratively.
At each iteration, the trend and variance estimation bandwidths are updated
by calling \href{h.cv.npf}{\code{h.cv}}.
}
\examples{
fd <- npf.data(ozone, dimnames = "day")
fit <- npf.fit(fd, maxlag = 100, var.h = 30, verbose = TRUE)
plot(fit)
}
\seealso{
\code{\link{npf.model}}, \code{\link{locpol.npf.bin}}, \code{\link{np.var.npf.bin.res2}},
\code{\link{np.svar.npf.locpol}}, \code{\link{h.cv.npf.bin}}, \code{\link{h.cv.npf.bin.res2}}, \code{\link[npsp:fitsvar.sb.iso]{npsp::fitsvar.sb.iso}}.
}
