% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npf.fit.R
\name{npf.model}
\alias{npf.model}
\alias{npf.model-class}
\alias{residuals.npf.model}
\alias{plot.npf.model}
\title{Nonparametric functional model}
\usage{
npf.model(lp, var, svm)

\method{residuals}{npf.model}(object, as.npf.data = TRUE, ...)

\method{plot}{npf.model}(
  x,
  y = NULL,
  main.trend = "Functional trend estimate",
  main.svar = "Semivariogram estimate",
  ...
)
}
\arguments{
\item{lp}{local polynomial estimate of the functional variance (object of
class \code{\link{npf.var}}).}

\item{var}{(optional) a vector or an object of class \code{\link{npf.var}} with the estimated (or
theoretical) variances.}

\item{svm}{fitted variogram model (object of class \code{\link[npsp:fitsvar.sb.iso]{npsp::fitsvar}}).}

\item{object}{object used to select a method.}

\item{as.npf.data}{logical; if \code{TRUE} (default) returns a \code{npf.data}-class
object containing the residuals.}

\item{...}{additional graphical parameters (passed to \code{\link{plot.npf.locpol}}).}

\item{x}{a nonparametric functional model object.
Typically an output of \code{\link{npf.fit}}.}

\item{y}{ignored argument.}

\item{main.trend}{title for the trend plot.}

\item{main.svar}{title for the semivariogram plot.}
}
\value{
Returns an object of \code{\link{class}} \code{npf.model}
(extending \code{\link{npf.locpol}}), the \code{lp} argument with
the others as additional components.
}
\description{
Defines a nonparametric functional model.
Constructor function of the S3 class \code{npf.model}.
}
\seealso{
\code{\link{npf.fit}}, \code{\link{npf.locpol}}, \code{\link{npf.var}}, \code{\link[npsp:fitsvar.sb.iso]{npsp::fitsvar.sb.iso}}.
}
