% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfda_internals.R
\name{npfda-internals}
\alias{npfda-internals}
\alias{is.fd1d}
\title{npfda internal and secondary functions}
\usage{
is.fd1d(object)
}
\description{
Listed below are supporting functions for the major methods in npfda.
}
\keyword{internal}
