% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfda_package.R
\docType{package}
\name{npfda-package}
\alias{npfda-package}
\alias{npfda}
\title{npfda: Nonparametric functional data analysis}
\description{
This package implements nonparametric methods for inference
on (multidimensional) functional data using the tools available in package
\href{https://rubenfcasal.github.io/npsp}{\code{npsp}}.
}
\details{
For more information visit \url{https://rubenfcasal.github.io/npfda.html}.
}
\references{
Fernández-Casal R. (2023) npsp: Nonparametric spatial (geo)statistics.
R package version 0.7-11, \url{https://rubenfcasal.github.io/npsp}.

Fernández-Casal R., Castillo-Páez S. and García-Soidán P. (2017), Nonparametric
estimation of the small-scale variability of heteroscedastic spatial processes,
\emph{Spa. Sta.}, \strong{22}, 358-370, \href{https://doi.org/10.1016/j.spasta.2017.04.001}{DOI}.
}
\keyword{Monte-Carlo}
\keyword{bootstrap}
\keyword{nonparametric}
\keyword{spatial}
