% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfda_package.R
\docType{data}
\name{ozone}
\alias{ozone}
\title{Ground-level ozone}
\format{
A matrix with 365 rows (days) and 33 columns (years).
}
\source{
The UK Automatic Urban and Rural Network (\href{https://uk-air.defra.gov.uk/networks/network-info?view=aurn}{AURN}).
The data were downloaded from \url{https://uk-air.defra.gov.uk/data} and preprocessed
with the \href{https://climatol.eu/}{\code{climatol}} package (by applying the usual outlier
detection and data imputation methods).
}
\description{
Daily averages of ozone concentration (microgram per cubic meter) recorded over
the period from 1988 to 2020 at the Yarner Wood \href{https://uk-air.defra.gov.uk/networks/network-info?view=aurn}{AURN}
monitoring site in the UK.
}
\details{
It is assumed that the observations corresponding to each year are (partial)
realizations of a functional process, so the data consist of 33 curves observed
at 365 discretization points.
}
\examples{
fd.ozone <- npf.data(ozone, dimnames = "day")
plot(fd.ozone)
}
\keyword{datasets}
