% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npf.data.R
\name{plot.npf.data}
\alias{plot.npf.data}
\title{Plot functional data}
\usage{
\method{plot}{npf.data}(
  x,
  y = seq_len(x$ny),
  slim = range(y, finite = TRUE),
  col = rainbow(128, end = 0.75),
  breaks = NULL,
  legend = TRUE,
  horizontal = FALSE,
  legend.shrink = 1,
  legend.width = 1.2,
  legend.mar = ifelse(horizontal, 3.1, 5.1),
  legend.lab = NULL,
  bigplot = NULL,
  smallplot = NULL,
  lab.breaks = NULL,
  axis.args = NULL,
  legend.args = NULL,
  add = FALSE,
  reset = TRUE,
  lty = 1,
  xlab = x$grid$dimnames,
  ylab = x$ylabel,
  ...
)
}
\arguments{
\item{x}{funcional data (\link{npf.data}-\link{class} object).}

\item{y}{numerical vector containing the values used for coloring the curves.}

\item{slim}{limits used to set up the color scale.}

\item{col}{color table used to set up the color scale (see \code{\link{image}} for
details).}

\item{breaks}{(optional) numeric vector with the breakpoints for the color scale: 
must have one more breakpoint than \code{col} and be in increasing order.}

\item{legend}{logical; if \code{TRUE} (default), the plotting region is splitted into two parts,
drawing the main plot in one and the legend with the color scale in the other.
If \code{FALSE} only the (coloured) main plot is drawn and the arguments related
to the legend are ignored (\code{\link[npsp:splot]{npsp::splot()}} is not called).}

\item{horizontal}{logical; if \code{FALSE} (default) legend will be a vertical strip on the
right side. If \code{TRUE} the legend strip will be along the bottom.}

\item{legend.shrink}{amount to shrink the size of legend relative to the
full height or width of the plot.}

\item{legend.width}{width in characters of the legend strip. Default is
1.2, a little bigger that the width of a character.}

\item{legend.mar}{width in characters of legend margin that has the axis.
Default is 5.1 for a vertical legend and 3.1 for a horizontal legend.}

\item{legend.lab}{label for the axis of the color legend. Default is no
label as this is usual evident from the plot title.}

\item{bigplot}{plot coordinates for main plot. If not passed, and \code{legend}
is \code{TRUE}, these will be determined within the function.}

\item{smallplot}{plot coordinates for legend strip. If not passed, and \code{legend}
is \code{TRUE}, these will be determined within the function.}

\item{lab.breaks}{if breaks are supplied these are text string labels to
put at each break value. This is intended to label axis on a transformed
scale such as logs.}

\item{axis.args}{additional arguments for the axis function used to create
the legend axis (see \code{\link[fields]{image.plot}} for details).}

\item{legend.args}{arguments for a complete specification of the legend
label. This is in the form of list and is just passed to the \code{\link{mtext}}
function. Usually this will not be needed (see \code{\link[fields]{image.plot}} 
for details).}

\item{add}{logical; if \code{TRUE} the plot is just added
to the existing plot.}

\item{reset}{logical; if \code{FALSE} the plotting region
(\code{par("plt")}) will not be reset to make it possible to add more features
to the plot (e.g. using functions such as points or lines). If \code{TRUE} (default)
the plot parameters will be reset to the values before entering the function.}

\item{lty}{vector of line types.}

\item{xlab}{label for the x axis, defaults to \code{x$grid$dimnames}.}

\item{ylab}{label for the y axis, defaults to \code{x$ylabel}.}

\item{...}{additional graphical parameters (to be passed to \code{\link[=matplot]{matplot()}}
or \code{\link[=matlines]{matlines()}}.
NOTE: graphical arguments passed here will only have impact on the main plot.
To change the graphical defaults for the legend use the \code{\link[=par]{par()}}
function beforehand (e.g. \code{par(cex.lab = 2)} to increase colorbar labels).}
}
\description{
Plots a functional data object.
}
\examples{
fd <- npf.data(ozone, dimnames = "day")
plot(fd)
plot(fd, y = as.numeric(fd$ynames))
}
\seealso{
\code{\link[npsp:splot]{npsp::splot()}}, \code{\link[=matplot]{matplot()}}, \code{\link[=matlines]{matlines()}}
}
\keyword{hplot}
