% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.svar.npf.R
\name{svar.bin.npf.data}
\alias{svar.bin.npf.data}
\alias{svar.bin.npf}
\alias{svar.bin.npf.locpol}
\title{Empirical (intra-curve) variogram estimate}
\usage{
\method{svar.bin}{npf.data}(
  x,
  maxlag = NULL,
  nlags = round(0.55 * x$grid$n),
  minlag = x$grid$lag,
  estimator = c("classical", "modulus"),
  ...
)

\method{svar.bin}{npf.locpol}(
  x,
  var,
  maxlag = NULL,
  nlags,
  minlag,
  estimator = c("classical", "modulus"),
  ...
)
}
\arguments{
\item{x}{object used to select a method.}

\item{maxlag}{maximum lag. Defaults to \code{(nlags - 1) * minlag}.}

\item{nlags}{number of lags. Defaults to \code{round(0.55 * x$grid$n)}.}

\item{minlag}{minimun lag. Defaults to that of the discretization grid.}

\item{estimator}{character, estimator name (e.g. \code{"classical"}).
See \code{\link[npsp:svar.bin]{npsp::svar.bin}}.}

\item{...}{further arguments passed to \code{\link[npsp:svar.bin]{npsp::svariso}}.}

\item{var}{(optional) a vector or an object of class \code{\link{npf.var}} with the estimated (or
theoretical) variances.}
}
\value{
Returns an object of S3 class \code{\link[npsp:svar.bin]{npsp::svar.bin}}.
}
\description{
Creates a \code{svar.bin} (binned semivar. + grid parameters) object with
the binned intra-curves semivariances (i.e. computes a binned sample variogram).
}
\details{
Only one of the arguments \code{maxlag} or \code{nlags} must be specified.

\code{svar.bin.npf.data} assumes a stationary functional process (the functional mean and
functional variance are constants).

See \code{\link[npsp:svar.bin]{npsp::svar.bin}} for more details.
}
\seealso{
\code{\link[npsp:svar.bin]{npsp::svar.bin}}, \code{\link{np.svar.npf.data}}, \code{\link{np.svar.npf.locpol}}
}
